requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-jazzmin==3.0.0
django-timezone-field==6.1.0
djangorestframework==3.15.2
kombu==5.3.7
prompt_toolkit==3.0.47
python-crontab==3.1.0
python-dateutil==2.9.0.post0
redis==5.0.6
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
